IS-3335: Refactor
