+++ src/sider/behandlerdialog/meldinger/Meldinger.tsx
@@ -1,13 +1,13 @@
 import React from "react";
 import { Box, Heading } from "@navikt/ds-react";
-import { Samtaler } from "@/sider/behandlerdialog/meldinger/Samtaler";
+import Samtaler from "@/sider/behandlerdialog/meldinger/Samtaler";
 import BehandleBehandlerdialogSvarOppgaveKnapp from "@/sider/behandlerdialog/meldinger/BehandleBehandlerdialogSvarOppgaveKnapp";
 
 export const texts = {
   header: "Meldinger",
 };
 
-export const Meldinger = () => {
+export default function Meldinger() {
   return (
     <Box background="surface-default" className="p-4">
       <Heading level="2" size="medium" spacing>
@@ -17,4 +17,4 @@ export const Meldinger = () => {
       <Samtaler />
     </Box>
   );
-};
+}
