+++ src/sider/behandlerdialog/meldinger/paminnelse/PaminnelseWarningIcon.tsx
